Fix monitoring page issues and enforce memory threshold minimum
- Fix initial chart time range not matching dropdown selection
- Fix 30-minute time range showing 1 hour data (float vs int parsing)
- Change update rate dropdown: 2 seconds → 3 seconds
- Convert all time displays from AM/PM to 24-hour format
- Enforce minimum 12GB threshold for large heap optimizations
- Remove unused security config settings (session cookies, lifetime)
- Add validation and warning when threshold set below 12GB

+++ routes.py
@@ -549,7 +549,7 @@ def get_metric_history(metric_type):
         from services.metrics_storage import metrics_storage
         
         # Get time range parameter (default to 1 hour)
-        hours = request.args.get('hours', 1, type=int)
+        hours = request.args.get('hours', 1, type=float)
         if hours <= 0 or hours > 168:  # Limit to 1 week max
             hours = 1
